XIVY-13806: Improve Tabs for WS, REST, DB
