+++ integrations/standalone/tests/integration/parts/index.ts
@@ -19,8 +19,10 @@ export * from './task';
 export * from './trigger';
 export * from './web-service';
 export * from './ws-request';
-export * from './ws-response';
-export * from './rest-response';
+export * from './ws-output';
+export * from './ws-error';
+export * from './rest-error';
+export * from './rest-output';
 export * from './rest-request';
 export * from './rest-request-body';
 export * from './permissions';
