+++ integrations/standalone/tests/integration/parts/ws-output.ts
@@ -1,0 +1,44 @@
+import type { Part } from '../../pageobjects/Part';
+import { NewPartTest, PartObject } from './part-tester';
+import type { Section } from '../../pageobjects/Section';
+import type { Table } from '../../pageobjects/Table';
+import type { ScriptArea } from '../../pageobjects/CodeEditor';
+
+class WsOutput extends PartObject {
+  mappingSection: Section;
+  mapping: Table;
+  codeSection: Section;
+  code: ScriptArea;
+
+  constructor(part: Part) {
+    super(part);
+    this.mappingSection = part.section('Mapping');
+    this.mapping = this.mappingSection.table(['label', 'expression']);
+    this.codeSection = part.section('Code');
+    this.code = this.codeSection.scriptArea();
+  }
+
+  async fill() {
+    await this.mappingSection.open();
+    await this.mapping.row(1).column(1).fill('"bla"');
+    await this.codeSection.open();
+    await this.code.fill('code');
+  }
+
+  async assertFill() {
+    await this.mapping.row(1).column(1).expectValue('"bla"');
+    await this.code.expectValue('code');
+  }
+
+  async clear() {
+    await this.mapping.row(1).column(1).clearExpression();
+    await this.code.clear();
+  }
+
+  async assertClear() {
+    await this.mapping.row(1).column(1).expectValue('');
+    await this.codeSection.expectIsClosed();
+  }
+}
+
+export const WsOutputTest = new NewPartTest('Output Data', (part: Part) => new WsOutput(part));
