+++ packages/editor/src/components/parts/index.ts
@@ -21,9 +21,10 @@ export * from './error/ErrorThrowPart';
 export * from './query/QueryPart';
 export * from './cache/CachePart';
 export * from './ws-request/WsRequestPart';
-export * from './ws-response/WsResponsePart';
+export * from './ws-error/WsErrorPart';
 export * from './web-service/WebServicePart';
-export * from './rest/RestResponsePart';
+export * from './rest/RestOutputPart';
+export * from './rest/RestErrorPart';
 export * from './rest/RestRequestPart';
 export * from './program/event/ProgramStartPart';
 export * from './program/activity/ProgramInterfaceStartPart';
