+++ packages/editor/src/components/parts/rest/useRestOutputData.ts
@@ -1,0 +1,33 @@
+import type { RestResponseData } from '@axonivy/inscription-protocol';
+import { produce } from 'immer';
+import type { DataUpdater } from '../../../types/lambda';
+import type { ConfigDataContext } from '../../../context';
+import { useConfigDataContext } from '../../../context';
+
+export function useRestOutputData(): ConfigDataContext<RestResponseData> & {
+  updateEntity: DataUpdater<RestResponseData['response']['entity']>;
+  resetData: () => void;
+} {
+  const { setConfig, ...config } = useConfigDataContext();
+
+  const updateEntity: DataUpdater<RestResponseData['response']['entity']> = (field, value) => {
+    setConfig(
+      produce(draft => {
+        draft.response.entity[field] = value;
+      })
+    );
+  };
+
+  const resetData = () =>
+    setConfig(
+      produce(draft => {
+        draft.response.entity = config.initConfig.response.entity;
+      })
+    );
+
+  return {
+    ...config,
+    updateEntity,
+    resetData
+  };
+}
