+++ packages/protocol/src/data/part-data.ts
@@ -43,7 +43,7 @@ export type WebserviceStartData = Pick<ConfigData, 'permission' | 'exception'>;
 
 export type WsRequestData = Pick<ConfigData, 'clientId' | 'operation' | 'properties'>;
 
-export type WsResponseData = Pick<ConfigData, 'output' | 'exceptionHandler'>;
+export type WsErrorData = Pick<ConfigData, 'exceptionHandler'>;
 
 export type ProcessDataData = Pick<ConfigData, 'data'>;
